requirements.txt
opencv-python>=4.8.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
mangum>=0.17.0
